+++ internal/config/config.go
@@ -80,7 +80,13 @@ func Load(file io.Reader) error {
 
 // Validate checks the current Config for sanity.
 func Validate() error {
-	for _, err := range []error{validateStorages(), validateToken(), SetGitPath(), validateShell()} {
+	for _, err := range []error{
+		validateStorages(),
+		validateToken(),
+		SetGitPath(),
+		validateShell(),
+		validateRuby(),
+	} {
 		if err != nil {
 			return err
 		}
@@ -94,6 +100,14 @@ func validateShell() error {
 	}
 
 	return validateIsDirectory(Config.GitlabShell.Dir, "gitlab-shell.dir")
+}
+
+func validateRuby() error {
+	if len(Config.Ruby.Dir) == 0 {
+		return fmt.Errorf("gitaly-ruby.dir is not set")
+	}
+
+	return validateIsDirectory(Config.Ruby.Dir, "gitaly-ruby.dir")
 }
 
 func validateIsDirectory(path, name string) error {